[chore] Fix doc and clippy error

+++ modules/axsignal/src/ucontext/mod.rs
@@ -5,7 +5,7 @@ pub const SS_ONSTACK: u32 = 1;
 /// The alternate signal stack is disabled.
 pub const SS_DISABLE: u32 = 2;
 /// The alternate signal stack has been marked to be autodisarmed as described above.
-pub const SS_AUTODISARM: u32 = (1 as u32) << 31;
+pub const SS_AUTODISARM: u32 = 1_u32 << 31;
 
 cfg_if::cfg_if! {
     if #[cfg(target_arch = "x86_64")] {
